This is going to be ridiculous
So, genius me failed to do the basic thing of RTFM of the Chrome Extension API
and learn about how the different runtimes are seggregated but the DOM is
shared. When you're trying to get at the page's runtime and you're stuck in your
own and the global variable you want differs, things get desperate.

So I tried to work around this rubbish and to be honest like a goat ramming a
dam I might just get there (and if I did, I'm sure there's bounty from Google in
it). Except it's getting late and I have better things to do. So what is this
abomination? Well, I can reach the DOM which means I can inject and execute
abitrary Javascript inside that DOM.

HA! TAKE THAT!

This is only going to get worse.