Adding button thing
I have a beef with this image because after pngcrush flattened it like a semi
trailer hitting wildlife tells me that a few percentages saved was image data
related and over half of it was cut down by simply removing the XMP metadata.

Why on earth is their more XMP metadata in an image file than actual image data?